+++ bin/describe/describe.ml
@@ -12,6 +12,9 @@ let subcommands =
   ; Describe_external_lib_deps.command
   ; Describe_opam_files.command
   ; Describe_pp.command
+  ; Printenv.command
+  ; Aliases_targets.Targets_cmd.command
+  ; Aliases_targets.Aliases_cmd.command
   ]
 
 let group =