+++ bin/printenv.mli
@@ -1,3 +1,5 @@
 open Import
 
+val term : unit Term.t
+
 val command : unit Cmd.t